// fe/src/lib/api/generated/model/graduateTracerDetailResponse.ts
export interface GraduateTracerDetailResponse {
  advanced_study_other_reason?: string;
  advanced_study_reasons?: GTSAdvancedStudyReasonValue[];
  birth_date?: string | null;
  civil_status?: GTSCivilStatusValue | null;
  created_at: string;
  current_employment_state?: GTSEmploymentStateValue | null;
  current_job_level?: GTSJobLevelValue | null;
  curriculum_improvement_suggestions?: string;
  curriculum_relevant_to_first_job?: boolean | null;
  degree_other_reason?: string;
  education?: GraduateTracerEducationPayload[];
  email?: string;
  employer_business_line?: GTSBusinessLineValue | null;
  first_job_after_college?: boolean | null;
  first_job_duration?: GTSFirstJobDurationValue | null;
  first_job_duration_other?: string;
  first_job_level?: GTSJobLevelValue | null;
  first_job_related_to_course?: boolean | null;
  first_job_source?: GTSFirstJobSourceValue | null;
  first_job_source_other?: string;
  graduate_study_reasons?: GTSDegreeReasonValue[];
  id: string;
  initial_gross_monthly_earning?: GTSEarningBracketValue | null;
  instrument_schema_version: number;
  mobile_number?: string;
  name?: string;
  permanent_address?: string;
  place_of_work?: GTSPlaceOfWorkValue | null;
  present_employment_status?: GTSPresentEmploymentStatusValue | null;
  present_occupation?: string;
  professional_exams?: GraduateTracerProfessionalExamPayload[];
  province?: string;
  reasons_for_accepting_first_job?: GTSJobReasonValue[];
  reasons_for_accepting_other?: string;
  reasons_for_changing_job?: GTSJobReasonValue[];
  reasons_for_changing_other?: string;
  reasons_for_staying_on_job?: GTSStayingReasonValue[];
  reasons_for_staying_other?: string;
  region_of_origin?: GTSRegionOfOriginValue | null;
  residence_location?: GTSResidenceLocationValue | null;
  self_employed_college_skills?: string;
  sex?: GTSSexValue | null;
  status: GraduateTracerStatusValue;
  student_id: string;
  submitted_at: string | null;
  telephone_contact_numbers?: string;
  time_to_first_job?: GTSFirstJobDurationValue | null;
  time_to_first_job_other?: string;
  trainings?: GraduateTracerTrainingPayload[];
  undergraduate_degree_reasons?: GTSDegreeReasonValue[];
  unemployment_other_reason?: string;
  unemployment_reasons?: GTSUnemploymentReasonValue[];
  updated_at: string;
  useful_competencies?: GTSUsefulCompetencyValue[];
  useful_competencies_other?: string;
}