// fe/src/lib/api/generated/model/graduateTracerEducationPayload.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface GraduateTracerEducationPayload {
  college_or_university: string;
  degree_and_specialization: string;
  honors_or_awards?: string;
  /** @minimum 1900 */
  year_graduated: number;
}