// fe/src/lib/api/generated/model/gTSAdvancedStudyReasonValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type GTSAdvancedStudyReasonValue = typeof GTSAdvancedStudyReasonValue[keyof typeof GTSAdvancedStudyReasonValue];


export const GTSAdvancedStudyReasonValue = {
  PROMOTION: 'PROMOTION',
  PROFESSIONAL_DEVELOPMENT: 'PROFESSIONAL_DEVELOPMENT',
  OTHER: 'OTHER',
} as const;