// fe/src/lib/api/generated/model/gTSCivilStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type GTSCivilStatusValue = typeof GTSCivilStatusValue[keyof typeof GTSCivilStatusValue];


export const GTSCivilStatusValue = {
  SINGLE: 'SINGLE',
  SEPARATED: 'SEPARATED',
  WIDOW_WIDOWER: 'WIDOW_WIDOWER',
  MARRIED: 'MARRIED',
  SINGLE_PARENT: 'SINGLE_PARENT',
} as const;